fix multiple deffinition in error handling
