+++ cpp/oneapi/dal/algo/rbf_kernel/backend/cpu/compute_kernel_dense.cpp
@@ -18,6 +18,7 @@
 
 #include "oneapi/dal/algo/rbf_kernel/backend/cpu/compute_kernel.hpp"
 #include "oneapi/dal/backend/interop/common.hpp"
+#include "oneapi/dal/backend/interop/error_converter.hpp"
 #include "oneapi/dal/backend/interop/table_conversion.hpp"
 
 namespace oneapi::dal::rbf_kernel::backend {
@@ -52,11 +53,12 @@ static compute_result call_daal_kernel(const context_cpu& ctx,
 
     daal_rbf_kernel::Parameter daal_parameter(desc.get_sigma());
 
-    interop::call_daal_kernel<Float, daal_rbf_kernel_t>(ctx,
-                                                        daal_x.get(),
-                                                        daal_y.get(),
-                                                        daal_values.get(),
-                                                        &daal_parameter);
+    interop::status_to_exception(
+        interop::call_daal_kernel<Float, daal_rbf_kernel_t>(ctx,
+                                                            daal_x.get(),
+                                                            daal_y.get(),
+                                                            daal_values.get(),
+                                                            &daal_parameter));
 
     return compute_result().set_values(
         homogen_table_builder{}.reset(arr_values, row_count_x, row_count_y).build());
